move code around outputting of statistics metadata to a file
