+++ urbanstats/games/quiz.py
@@ -13,21 +13,18 @@ import pytz
 import tqdm.auto as tqdm
 from permacache import permacache, stable_hash
 
-from create_website import (
-    get_index_lists,
-    shapefile_without_ordinals,
-    statistic_internal_to_display_name,
-)
-from produce_html_page import (
-    create_filename,
-    get_statistic_categories,
-    indices,
-    internal_statistic_names,
-)
+from create_website import shapefile_without_ordinals
+from produce_html_page import create_filename, indices
 from relationship import states_for_all
 from shapefiles import american_to_international, filter_table_for_type
 from urbanstats.shortener import shorten
 from urbanstats.statistics.collections_list import statistic_collections
+from urbanstats.statistics.output_statistics_metadata import (
+    get_statistic_categories,
+    internal_statistic_names,
+    statistic_internal_to_display_name,
+    get_statistic_column_path,
+)
 
 from .fixed import juxtastat as fixed_up_to
 
@@ -315,7 +312,6 @@ def generate_quizzes(folder):
 
 
 def generate_quiz_info_for_website(site_folder):
-    from create_website import get_statistic_column_path
 
     folder = "react/src/data/quiz"
     try:
@@ -491,10 +487,8 @@ for collection in statistic_collections:
 stats = sorted(stats_to_display, key=str)
 categories = sorted({get_statistic_categories()[x] for x in stats})
 
-unrecognized = (set(stats) | set(not_included)) - set(
-    statistic_internal_to_display_name()
-)
+unrecognized = (set(stats) | set(not_included)) - set(internal_statistic_names())
 assert not unrecognized, unrecognized
 
-extras = set(statistic_internal_to_display_name()) - (set(stats) | set(not_included))
+extras = set(internal_statistic_names()) - (set(stats) | set(not_included))
 assert not extras, extras
